+++ my_music_code/lib/Album/album_page.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:my_music_code/Feed/Components/feed_music_grid.dart';
+import 'package:my_music_code/Models/music_model.dart';
 import 'package:my_music_code/Music/music_page.dart';
 import 'package:my_music_code/Globals/responsive_container.dart';
 import 'package:my_music_code/Globals/responsive_text.dart';
