+++ my_music_code/lib/Feed/Components/feed_horizontal_scroll_component.dart
@@ -6,16 +6,6 @@ import 'package:my_music_code/Globals/responsive_text.dart';
 import 'package:my_music_code/Globals/size_config.dart';
 import 'package:my_music_code/Globals/style.dart';
 
-// class AlbumBase {
-//   final String name;
-//   final String id;
-//   final String artist;
-//   final String imageUrl;
-//   final List<Music> listMusic;
-
-//   AlbumBase({required this.name, required this.id, required this.artist, required this.imageUrl, required this.listMusic});
-// }
-
 class FeedHorizontalScrollComponent extends StatefulWidget {
   const FeedHorizontalScrollComponent({super.key, required this.title, required this.albuns});
   final String title;
