+++ my_music_code/lib/Feed/Components/new_music_release.dart
@@ -1,6 +1,6 @@
 
 import 'package:flutter/material.dart';
-import 'package:my_music_code/Feed/Components/feed_music_grid.dart';
+import 'package:my_music_code/Models/music_model.dart';
 import 'package:my_music_code/Music/music_page.dart';
 import 'package:my_music_code/Globals/responsive_container.dart';
 import 'package:my_music_code/Globals/responsive_text.dart';
