+++ my_music_code/lib/Globals/navigator_page.dart
@@ -3,9 +3,9 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:my_music_code/Album/album_page.dart';
-import 'package:my_music_code/Feed/Components/feed_music_grid.dart';
 import 'package:my_music_code/Feed/feed_page.dart';
 import 'package:my_music_code/Globals/mini_player.dart';
+import 'package:my_music_code/Models/music_model.dart';
 import 'package:my_music_code/MyPlaylists/user_page_of_playlists.dart';
 import 'package:my_music_code/Search/search_page.dart';
 import 'package:my_music_code/Globals/style.dart';
