+++ my_music_code/lib/Models/music_model.dart
@@ -1,0 +1,10 @@
+class Music {
+  final String? name;
+  final String? id;
+  final String? artist;
+  final String? imageUrl;
+  final String? link;
+  final int? duration;
+
+  Music({this.name, this.id, this.artist, this.imageUrl, this.link, this.duration});
+}
